test: Move to TypeScript
