+++ package.json
@@ -37,6 +37,7 @@
     "universal-user-agent": "^3.0.0"
   },
   "devDependencies": {
+    "@types/fetch-mock": "^7.2.5",
     "@types/jest": "^24.0.13",
     "@types/node": "^12.0.2",
     "compression-webpack-plugin": "^2.0.0",
@@ -48,6 +49,7 @@
     "semantic-release": "^15.13.3",
     "simple-mock": "^0.8.0",
     "standard": "^13.0.1",
+    "ts-jest": "^24.0.2",
     "tslint": "^5.16.0",
     "tslint-config-prettier": "^1.18.0",
     "tslint-config-standard": "^8.0.1",
@@ -87,5 +89,8 @@
   },
   "files": [
     "lib"
-  ]
+  ],
+  "jest": {
+    "preset": "ts-jest"
+  }
 }
